move dialog file
