+++ frontend/packages/design-system/src/components/gantt-view/deleteAllocationDialog.tsx
